Spelling Fix to index.js in Questions
Fixed minor spelling mistakes and added a small header below the result statement to suggest the user to visit the resources page. Also if the result is the Peace and Freedom Party, the page shows The Peace Party, which can be confusing. I noticed that the variable used to track Peace And Freedom responses is called peace, so maybe you can add an if statement that sets result to Peace And Freedom if peace will be the result. I'd do it myself but I'm not very experienced in JS.